requirements.txt
requests
pandas
python-dotenv
tqdm
matplotlib
seaborn
jupyter
ipykernel